+++ packages/core/src/index.ts
@@ -1,38 +1,11 @@
 import Buffer from 'buffer'
 import CID from 'cids'
 
-export { ID, Variant } from './thread/id'
-export { Multiaddr, PrivateKey, PublicKey, Info, LogInfo } from './thread/info'
+export * from './thread'
+export * from './service'
+export * from './external'
 
 export interface Block {
   data: Buffer
   cid: CID
-}
-
-export interface BaseRecord {
-  event_node?: string
-  header_node?: string
-  body_node?: string
-  record_node?: string
-}
-export interface LogRecord extends BaseRecord {
-  id: string
-  log_id?: string
-  thread_id: string
-}
-
-export interface HeaderNode {
-  key: Buffer
-  time: number
-}
-
-export interface EventNode {
-  header: any
-  body: any
-}
-
-export interface RecordNode {
-  sig: Buffer
-  prev: any
-  block: any
 }
