+++ packages/core/src/service/index.ts
@@ -1,0 +1,14 @@
+import CID from 'cids'
+import { LogID, ThreadID } from '../thread'
+
+export interface ThreadRecord {
+  value: RecordNode
+  threadID: ThreadID
+  logID: LogID
+}
+
+export interface RecordNode {
+  sig: Buffer
+  prev: CID
+  block: CID
+}
