+++ packages/core/src/thread/index.ts
@@ -1,0 +1,2 @@
+export * from './id'
+export * from './info'
