+++ packages/encoding/tsconfig.json
@@ -6,6 +6,7 @@
   },
   "exclude": [
     "node_modules",
-    "dist"
+    "dist",
+    "**/*.spec.ts"
   ]
 }
