+++ packages/service/src/logstore/addrbook.spec.ts
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 import { expect } from 'chai'
 import { MemoryDatastore } from 'interface-datastore'
-import { ID, Variant } from '@textile/threads-core'
+import { ThreadID, Variant } from '@textile/threads-core'
 
 import { AddrBook } from './addrbook'
 
@@ -24,12 +24,12 @@ describe('AddrBook', () => {
       const mem = new MemoryDatastore()
       ab = new AddrBook(mem, { ttl: 100, frequency: 20 })
     })
-    const id = ID.fromRandom(Variant.Raw, 24)
+    const id = ThreadID.fromRandom(Variant.Raw, 24)
     it('add a single address', async () => {
       const log = 'peer id' + Math.random()
       const addr = generateAddrs(1)[0]
-      await ab.put(id, log, 1000 * 60 * 60, addr) // 1 hr
-      expect(await ab.get(id, log)).to.deep.equal([addr])
+      await ab.put(id, log, 1000, addr) // 1 hr
+      expect(await ab.get(id, log)).to.deep.equal(new Set([addr]))
     })
     it('idempotent add single address', async () => {
       const log = 'peer id' + Math.random()
@@ -56,10 +56,10 @@ describe('AddrBook', () => {
       const addr = generateAddrs(3)
       await ab.put(id, log, 200, ...addr) // 1 sec
       // same address as before but with a higher TTL
-      await ab.put(id, log, 200 * 60, addr[2]) // 1 min
+      await ab.put(id, log, 600, addr[2]) // 1 min
       // after the initial TTL has expired, check that only the third address is present.
-      await sleep(300)
-      expect(await ab.get(id, log)).to.have.length(1)
+      await sleep(400)
+      expect((await ab.get(id, log)).size).to.equal(1)
       // make sure we actually set the TTL
       await ab.put(id, log, 0)
       expect(await ab.get(id, log)).to.have.length(0)
@@ -74,7 +74,7 @@ describe('AddrBook', () => {
       // after the initial TTL has expired, check that all three addresses are still present (i.e. the TTL on
       // the modified one was not shortened)
       await sleep(210)
-      expect(await ab.get(id, log)).to.have.length(3)
+      expect((await ab.get(id, log)).size).to.equal(3)
     })
     it('adding an existing address with an earlier expiration never reduces the TTL', async () => {
       const log = 'peer id' + Math.random()
@@ -88,14 +88,14 @@ describe('AddrBook', () => {
       await sleep(200)
       // 1 unit left
       // We still have the address
-      expect(await ab.get(id, log)).to.have.length(1)
+      expect((await ab.get(id, log)).size).to.equal(1)
       // The TTL wasn't reduced
       await ab.put(id, log, 0)
-      expect(await ab.get(id, log)).to.have.length(0)
+      expect((await ab.get(id, log)).size).to.equal(0)
     })
   })
   describe('Clearing', () => {
-    const id = ID.fromRandom(Variant.Raw, 24)
+    const id = ThreadID.fromRandom(Variant.Raw, 24)
     beforeEach(async () => {
       ab = new AddrBook(new MemoryDatastore(), { ttl: 3600, frequency: 30 })
     })
@@ -104,15 +104,15 @@ describe('AddrBook', () => {
       const addr = generateAddrs(5)
       await ab.put(id, logs[0], 100 * 60, ...addr.slice(0, 3))
       await ab.put(id, logs[1], 100 * 60, ...addr.slice(3, 5))
-      expect(await ab.get(id, logs[0])).to.deep.equal(addr.slice(0, 3))
-      expect(await ab.get(id, logs[1])).to.deep.equal(addr.slice(3))
+      expect([...(await ab.get(id, logs[0]))]).to.deep.equal(addr.slice(0, 3))
+      expect([...(await ab.get(id, logs[1]))]).to.deep.equal(addr.slice(3))
       await ab.clear(id, logs[0])
 
       expect(await ab.get(id, logs[0])).to.have.length(0)
-      expect(await ab.get(id, logs[1])).to.deep.equal(addr.slice(3))
+      expect([...(await ab.get(id, logs[1]))]).to.deep.equal(addr.slice(3))
       await ab.clear(id, logs[1])
-      expect(await ab.get(id, logs[0])).to.have.length(0)
-      expect(await ab.get(id, logs[1])).to.have.length(0)
+      expect((await ab.get(id, logs[0])).size).to.equal(0)
+      expect((await ab.get(id, logs[1])).size).to.equal(0)
     })
   })
 })
