+++ packages/service/src/logstore/index.ts
@@ -1,0 +1,5 @@
+export { AddrBook } from './addrbook'
+export { HeadBook } from './headbook'
+export { KeyBook } from './keybook'
+export { MetadataBook } from './metadatabook'
+export { LogStore } from './logstore'
