+++ packages/service/src/logstore/interface.ts
@@ -1,0 +1,12 @@
+import { ThreadID, LogID } from '@textile/threads-core'
+
+export interface LogsThreads {
+  // logs returns a list of log IDs for a thread.
+  logs(id: ThreadID): Promise<Set<LogID>>
+  // threads returns a list of threads referenced in the book.
+  threads(): Promise<Set<ThreadID>>
+}
+
+export interface Closer {
+  close(): Promise<void>
+}
