+++ packages/service/src/logstore/keybook.spec.ts
@@ -1,15 +1,15 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
 import { expect } from 'chai'
 import { MemoryDatastore } from 'interface-datastore'
-import { ID, PrivateKey } from '@textile/threads-core'
+import { ThreadID, PrivateKey } from '@textile/threads-core'
 import { KeyBook } from './keybook'
 
+// @todo: Find or write type defs for these (or replace them with smaller deps)
 const crypto = require('libp2p-crypto')
 const PeerId = require('peer-id')
 
 let kb: KeyBook
-const tid: ID = ID.fromRandom(0, 24)
-const peer = new PeerId(Buffer.from('test peer'))
+const tid: ThreadID = ThreadID.fromRandom(0, 24)
 
 describe('KeyBook', () => {
   before(async () => {
@@ -20,59 +20,47 @@ describe('KeyBook', () => {
     await kb.close()
   })
   it('PrivKey', async () => {
-    const privKey: PrivateKey | undefined = await crypto.keys.generateKeyPair('rsa', 1024)
-    if (privKey === undefined) {
-      expect(privKey).to.not.be.undefined
-      return
-    }
+    const privKey: PrivateKey = await crypto.keys.generateKeyPair('rsa', 1024)
+    expect(privKey).to.not.be.undefined
 
     const log = await PeerId.createFromPrivKey(privKey.bytes)
     const logId = log.toB58String()
 
     // No privkey exists yet
-    try {
-      const err = await kb.privKey(tid, logId)
-      throw new Error('should have rejected')
-    } catch (err) {
-      expect(err.toString()).to.equal('Error: Not Found')
-    }
+    const key = await kb.privKey(tid, logId)
+    expect(key).to.be.undefined
 
     // Add should not err
     const res = await kb.addPrivKey(tid, logId, privKey)
-    expect(res).to.equal(undefined)
+    expect(res).to.be.undefined
 
     // Stored priv key should match
     const que = await kb.privKey(tid, logId)
-    expect(que.bytes).to.deep.equal(privKey.bytes)
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    expect(que!.bytes).to.deep.equal(privKey!.bytes)
 
     const logs = await kb.logs(tid)
     expect(Array.from(logs)).to.contain(logId)
   })
   it('PubKey', async () => {
-    const privKey: PrivateKey | undefined = await crypto.keys.generateKeyPair('rsa', 1024)
-    if (privKey === undefined) {
-      expect(privKey).to.not.be.undefined
-      return
-    }
+    const privKey: PrivateKey = await crypto.keys.generateKeyPair('rsa', 1024)
+    expect(privKey).to.not.be.undefined
 
     const log = await PeerId.createFromPubKey(privKey.public.bytes)
     const logId = log.toB58String()
 
     // No pubKey exists yet
-    try {
-      const err = await kb.pubKey(tid, logId)
-      throw new Error('should have rejected')
-    } catch (err) {
-      expect(err.toString()).to.equal('Error: Not Found')
-    }
+    const key = await kb.pubKey(tid, logId)
+    expect(key).to.be.undefined
 
     // Add should not err
     const res = await kb.addPubKey(tid, logId, privKey.public)
-    expect(res).to.equal(undefined)
+    expect(res).to.be.undefined
 
     // Stored priv key should match
     const que = await kb.pubKey(tid, logId)
-    expect(que.bytes).to.deep.equal(privKey.public.bytes)
+    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+    expect(que!.bytes).to.deep.equal(privKey!.public.bytes)
 
     const logs = await kb.logs(tid)
     expect(Array.from(logs)).to.contain(logId)
