--- packages/service/src/options.ts
@@ -1,21 +1,0 @@
-// import { Multiaddr } from 'multiaddr'
-// import { ID } from '../thread/id'
-
-// // AddOptions holds values used for an add operation.
-// export interface AddOptions {
-//   threadId?: ID
-//   time?: Date
-//   key?: Buffer
-//   keyLog?: string
-//   addrs?: Multiaddr[]
-// }
-
-// // PutOptions holds values used for a put operation.
-// export interface PutOptions {
-//   threadID?: ID
-//   logID?: string
-// }
-
-// export interface SubOptions {
-//   threadIDs?: ID[]
-// }
