+++ packages/utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@textile/threads-core",
+  "name": "@textile/threads-utils",
   "version": "0.0.1",
   "main": "dist/index",
   "types": "dist/index",
@@ -12,18 +12,15 @@
     "build": "tsc -p tsconfig.json",
     "clean": "rimraf ./dist"
   },
+  "dependencies": {
+    "@textile/threads-core": "0.0.1",
+    "cbor-sync": "^1.0.4",
+    "loglevel": "^1.6.6",
+    "multiaddr": "^7.2.1",
+    "multicodec": "0.5.6"
+  },
   "devDependencies": {
-    "@types/node": "^12.12.21",
-    "@types/randombytes": "^2.0.0",
-    "@types/varint": "^5.0.0",
-    "cids": "github:multiformats/js-cid#master",
     "rimraf": "^3.0.0",
     "typescript": "^3.7.2"
-  },
-  "dependencies": {
-    "loglevel": "^1.6.6",
-    "multibase": "^0.6.0",
-    "randombytes": "^2.1.0",
-    "varint": "^5.0.0"
   }
 }
