+++ packages/utils/src/index.ts
@@ -1,0 +1,5 @@
+import { PrivateKey } from '@textile/threads-core'
+
+function logFromPrivateKey(key: PrivateKey): string {
+  return 'log'
+}
